Implementation for first test scenario
This commit includes the definitions of the following classes:

- Card
- Deck
- Game
- Simulator
- Hand

These classes have been identified and modeled based on the description given
Simulator and Game serve as the classes responsible for setup up the basic
requirements for the blackjack. Each comprises of a play method that executes
the game.

Tests have been included that describes eac class and methods.